feat: implement post view recording with fingerprint generation
