+++ frontend/app/posts/[username]/[slug]/post-client.tsx
@@ -2,15 +2,11 @@
 import React, { useEffect, useState } from "react";
 import Link from "next/link";
 import { FaCalendar, FaClock, FaUser } from "react-icons/fa";
-import { IoChevronDown, IoChevronForward } from "react-icons/io5";
+import { IoChevronForward } from "react-icons/io5";
 import ScrollProgressBar from "@/app/components/ScrollProgress";
-import { PreviewEditor, DynamicPreviewEditor } from "@/app/components/tiptap-templates/simple/view-editor";
-import { Button } from "@/components/ui/button";
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
-import { AlignJustify } from "lucide-react";
+import { DynamicPreviewEditor } from "@/app/components/tiptap-templates/simple/view-editor";
 import { JSONContent } from "@tiptap/react";
 import { Post } from "@/app/interfaces";
-import { toast, usePostView } from "@/hooks/use-toast";
 import NotFound from "@/app/components/NotFound";
 import { SEOProvider } from "@/app/contexts/seoContext";
 import Loading from "@/app/components/Loading";
@@ -18,7 +14,6 @@ import BlogAIChat from "@/app/components/ai-chat";
 import { generateStructuredData } from '@/app/utils/seo';
 import Image from "next/image";
 import { useRouter, useSearchParams } from 'next/navigation';
-import { generateFingerprint } from "@/lib/utils";
 
 interface PostClientProps {
   post: Post;
@@ -33,7 +28,6 @@ export default function PostClient({ post, isLoadingPost }: PostClientProps) {
   const [isClient, setIsClient] = useState(false);
   const searchParams = useSearchParams();
   const router = useRouter();
-  const { recordView } = usePostView();
 
   const isChatOpen = searchParams.get('chat') === 'true';
   const isChatFullOpen = searchParams.get('chat_full') === 'true';
@@ -57,20 +51,6 @@ export default function PostClient({ post, isLoadingPost }: PostClientProps) {
     setIsClient(true);
   }, []);
 
-  // Record post view when post loads
-  useEffect(() => {
-    if (isClient && post?.id) {
-      const recordViewAsync = async () => {
-        try {
-          const fingerprint = await generateFingerprint();
-          recordView(post.id, fingerprint);
-        } catch (error) {
-          console.error('Failed to generate fingerprint:', error);
-        }
-      };
-      recordViewAsync();
-    }
-  }, [isClient, post?.id, recordView]);
 
   // Update metadata when post changes
   useEffect(() => {
